Encode a common pattern as the default behavior.
It is a common pattern when using GenRegistry to wrap lookup_or_start/4
with a call to lookup/2 first.

This pattern exists because it allows the caller to do a fast ETS
lookup in the client context without having to switch to the server
context when the process is already running.

lookup_or_start/4 accepts a GenServer.server(), in the cases that a
caller uses a pid, the old path will be followed since there's no good
way to go from pid to :ets.tab().  If the caller provides an atom
(local named process) then the atom will be assumed to be the
:ets.tab() and a client-side lookup performed.