Adds a start/4
start/4 allows the caller to know whether or not they started the
worker process.  lookup_or_start/4 is transparent about whether the
worker process was started fresh or already existed. Some use cases
require that the caller be able to tell whether or not they started
the worker process.

start/4 will error if the worker process already exists with
`{:error, {:already_started, pid}}` similar to how a Supervisor
behaves when a child is already running.